#55 define model searchUser and move function getAllUser into model user, fix bug function handleSearchUsers in controller
